move the abstract base tests back to api module
