+++ api/src/test/scala/com/heroku/platform/api/AddonSpec.scala
@@ -1,6 +1,4 @@
-package com.heroku.platform.api.client.spray
-
-import com.heroku.platform.api._
+package com.heroku.platform.api
 
 abstract class AddonSpec(aj: ApiRequestJson with ApiResponseJson) extends ApiSpec(aj) {
 
@@ -21,5 +19,4 @@ abstract class AddonSpec(aj: ApiRequestJson with ApiResponseJson) extends ApiSpe
   }
 
 }
-
