+++ api/src/test/scala/com/heroku/platform/api/AppSpec.scala
@@ -1,10 +1,6 @@
-package com.heroku.platform.api.client.spray
-
-import com.heroku.platform.api._
-
+package com.heroku.platform.api
 
 abstract class AppSpec(aj: ApiRequestJson with ApiResponseJson) extends ApiSpec(aj) {
-
 
   val implicits: HerokuAppRequestJson with HerokuAppResponseJson with ErrorResponseJson = aj
 
@@ -32,7 +28,6 @@ abstract class AppSpec(aj: ApiRequestJson with ApiResponseJson) extends ApiSpec(
       updated.name must be(newname)
 
       delete(HerokuApp.Delete(updated.id))
-
 
     }
   }
